+++ sdk/communication/AzureCommunicationChat/Source/ChatThreadClient.swift
@@ -319,18 +319,18 @@ public class ChatThreadClient {
     /// Updates a message.
     /// - Parameters:
     ///    - message: The message id.
-    ///    - content: The updated message content.
+    ///    - parameters: The updated message content.
     ///    - options: Update chat message options
     ///    - completionHandler: A completion handler that receives a status code on success.
     public func update(
         message messageId: String,
-        content: UpdateChatMessageRequest,
+        parameters: UpdateChatMessageRequest,
         withOptions options: UpdateChatMessageOptions? = nil,
         completionHandler: @escaping HTTPResultHandler<Void>
     ) {
         service
             .update(
-                chatMessage: content,
+                chatMessage: parameters,
                 chatThreadId: threadId,
                 chatMessageId: messageId,
                 withOptions: options