+++ sdk/communication/AzureCommunicationChat/Tests/ChatThreadClientUnitTests.swift
@@ -380,7 +380,7 @@ class ChatThreadClientUnitTests: XCTestCase {
         let expectation = self.expectation(description: "Update message")
 
         let message = UpdateChatMessageRequest(content: "update message")
-        chatThreadClient.update(message: messageId, content: message, completionHandler: { result, _ in
+        chatThreadClient.update(message: messageId, parameters: message, completionHandler: { result, _ in
             switch result {
             case let .success(response):
                 XCTAssertNotNil(response)